Added blanket implementation
BUG: rust analyzer does not like the two new blanket implementations  in
`src/layout.rs` when used in conjunction with the blanket
implementations	for references of `Widget` in `src/widget/mod.rs`. The
bug can be reproduced simply by creating one of the flex containers
containing some items. The code builds fine though.

It might be reasonable to hide these implementations behind a feature
flag.